do not write to static field from instance method
The code was created by the openhab binding template.
As long as there is not reason to call the previously static
getContext() function to get the context of the started OSGi bundle, we
keep that function as a non static member function.
If the loss of the static function will break something, we could add
the static keyword again.
A grep through the openHAB sources shows bindings that are using the
getContext() function (statically) in their code themself.
We are not using that.
Let's show, what's going on.